work on Diagrammatic
