+++ oalg-base/src/OAlg/Limes/Diagrammatic.hs
@@ -10,7 +10,7 @@
   , StandaloneDeriving
   , TypeOperators
   , DataKinds
-  , RankNTypes
+  , ConstraintKinds
 #-}
 
 
@@ -48,14 +48,13 @@ import Data.Typeable
 import OAlg.Prelude
 
 import OAlg.Data.Relation
+import OAlg.Data.Duality
 
 import OAlg.Hom.Oriented.Definition
 
 import OAlg.Structure.Oriented.Definition
 
 import OAlg.Entity.Diagram hiding (DiagramDuality(..))
-
-import OAlg.Limes.OpDuality
 
 --------------------------------------------------------------------------------
 -- Diagrammatic -
@@ -64,8 +63,7 @@ import OAlg.Limes.OpDuality
 class Diagrammatic d where
   diagram :: d t n m a -> Diagram t n m a
 
-instance Diagrammatic Diagram where
-  diagram = id
+instance Diagrammatic Diagram where diagram = id
 
 --------------------------------------------------------------------------------
 -- DiagrammaticApplicative -
@@ -74,11 +72,25 @@ instance Diagrammatic Diagram where
 --
 -- __Property__ Let @h@ be in @__h__ __a__ __b__@ for @'HomOriented' __h__@ and @'Diagrammatic' __d__@,
 -- then holds: For all @__t__@, @__n__@, @__m__@, @__a__@ and @d@ in @__d__ __t__ __n__ __m__ __a__@
--- holds: @'diagram' ('dmap' h d) '==' 'dgMap' h ('diagram' d)@.
+-- holds: @'diagram' ('dmap' h d) '==' 'dmap' h ('diagram' d)@.
 class (HomOriented h, Diagrammatic d) => DiagrammaticApplicative h d where
   dmap :: h a b -> d t n m a -> d t n m b
 
-instance HomOriented h => DiagrammaticApplicative h Diagram where dmap = dgMap
+instance HomOriented h => DiagrammaticApplicative h Diagram where dmap = amap1
+
+--------------------------------------------------------------------------------
+-- DiagrammaticApplicative1 -
+
+-- | applications on 'Diagrammatic' objects as a one-parameterized type, where 'dmap' and 'amap1'
+-- are equaly defined.
+--
+-- __Property__ Let @'DiagrammaticApplicative __h__ __d__@ and
+-- @'Applicative1' __h__ (__d__ __t__ __n__ __m__)@, then holds:
+-- @'dmap' h d '==' 'amap1' h d@ for all @__a__@, @__b__@, @h@ in @__h__ __a__ __b__@ and @d@ in
+-- @__d__ __t__ __n__ __m__ __a__@.
+class (DiagrammaticApplicative h d, Applicative1 h (d t n m)) => DiagrammaticApplicative1 h d t n m
+
+instance HomOriented h => DiagrammaticApplicative1 h Diagram t n m
 
 --------------------------------------------------------------------------------
 -- DiagrammaticFunctorial -
@@ -96,6 +108,26 @@ instance HomOriented h => DiagrammaticApplicative h Diagram where dmap = dgMap
 -- @d@ in @__d__ __t__ __n__ __m__ __x__@. 
 class (FunctorialHomOriented h, DiagrammaticApplicative h d) => DiagrammaticFunctorial h d
 
+--------------------------------------------------------------------------------
+-- DiagrammaticFunctorial1 -
+
+-- | functorial applications on 'Diagrammatic' objects as a one-parameterized type.
+--
+-- __Note__ Actually from  @'DiagrammaticApplicative1' __h__ __d__ __t__ __n__ __m__@ and
+-- @'Functorial1'  __h__ (__d__ __t__ __n__ __m__)@ it follows that
+-- @'DiagrammaticFunctorial __h__ __d__@ holds, but it is not feasible to declare an
+-- instance:
+--
+-- @
+-- instance (DiagrammaticApplicative1 h d t n m, Functorial1 h (d t n m))
+--   => DiagrammaticFunctorial h d
+-- @
+type DiagrammaticFunctorial1 h d t n m
+  = ( DiagrammaticApplicative1 h d t n m
+    , DiagrammaticFunctorial h d
+    , Functorial1 h (d t n m)
+    )
+    
 instance FunctorialHomOriented h => DiagrammaticFunctorial h Diagram
 
 --------------------------------------------------------------------------------
@@ -110,7 +142,6 @@ instance FunctorialHomOriented h => DiagrammaticFunctorial h Diagram
 -- @'Eq' (__d__ __t__ __n__ __m__ __a__)@.
 class ( Diagrammatic d
       , DiagrammaticFunctorial (IsoOp s) d
-      , TransformableOrt s, TransformableOp s, TransformableTyp s
       ) => DiagrammaticDualisable s d where
   coDiagrammatic :: Struct s a -> d t n m a -> d (Dual t) n m (Op a)
 
@@ -125,8 +156,8 @@ instance (TransformableOrt s, TransformableOp s, TransformableTyp s)
 data DiagrammaticDuality s d i o a b where
   DiagrammaticDuality
     :: ( DiagrammaticDualisable s d
-       , Functorial1 (IsoOp s) (d t n m)
-       , Functorial1 (IsoOp s) (d (Dual t) n m)
+       , DiagrammaticFunctorial1 (IsoOp s) d t n m
+       , DiagrammaticFunctorial1 (IsoOp s) d (Dual t) n m
        )
     => Dual (Dual t) :~: t
     -> DiagrammaticDuality s d (IsoOp s) Op (d t n m) (d (Dual t) n m)
@@ -135,77 +166,32 @@ instance Symmetric (DiagrammaticDuality s d i o) where
   swap (DiagrammaticDuality Refl) = DiagrammaticDuality Refl
 
 instance (TransformableTyp s, Transformable1 Op s)
-  => Duality1 s (DiagrammaticDuality s dy) (IsoOp s) Op where
+  => Duality1 s (DiagrammaticDuality s d) (IsoOp s) Op where
   toDual1 (DiagrammaticDuality _)    = coDiagrammatic
   isoBidual1 (DiagrammaticDuality _) = Functor1 . isoToOpOpStruct   
 
-{-
 --------------------------------------------------------------------------------
--- dgmdSoundDual1 -
+-- DiagrammaticOpDuality -
 
--- | soundness of 'Dual1' for 'DiagrammaticDuality's.
-dgmdSoundDual1 :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (d t n m) :~: d (Dual t) n m
-dgmdSoundDual1 (DiagrammaticDuality r) = diagrammaticSoundDual1 r
+-- | diagrammatic 'Op' duality.
+type DiagrammaticOpDuality s d t n m = DiagrammaticDuality s d (IsoOp s) Op (d t n m) (d (Dual t) n m)
 
 --------------------------------------------------------------------------------
--- DiagrammaticDuality - Duality1 -
+-- dgmOpDuality -
 
+-- | 'Op'-duality for 'Diagrams's.
+dgmOpDuality :: (TransformableOrt s, TransformableOp s, TransformableTyp s)
+  => Dual (Dual t) :~: t
+  -> DiagrammaticOpDuality s Diagram t n m
+dgmOpDuality = DiagrammaticDuality
 
-instance TransformableDual1 (DiagrammaticDuality s d i o) where
-  tauDual1 d@(DiagrammaticDuality _) = tauDual1' d
-
-tauDual1' :: DiagrammaticDuality s d i o (d t n m)
-  -> DiagrammaticDuality s d i o (Dual1 (d t n m))
-tauDual1' d@(DiagrammaticDuality _) = tauDual1'' d (dgmdSoundDual1 d)
-
-tauDual1'' :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (d t n m) :~: d (Dual t) n m
-  -> DiagrammaticDuality s d i o (Dual1 (d t n m))
-tauDual1'' (DiagrammaticDuality Refl) Refl = DiagrammaticDuality Refl
-
-
-instance ReflexiveDual1 (DiagrammaticDuality s d i o) where
-  reflDual1 d@(DiagrammaticDuality _) = reflDual1' d
-  
-reflDual1' :: DiagrammaticDuality s d i o (d t n m) -> Dual1 (Dual1 (d t n m)) :~: d t n m
-reflDual1' d = reflDual1'' d (lemma1 d) (lemma3 d)
-
-lemma1 :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (Dual1 (d t n m)) :~: Dual1 (d (Dual t) n m)
-lemma1 d = lemma2 d (dgmdSoundDual1 d)
-
-lemma2 :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (d t n m) :~: d (Dual t) n m
-  -> Dual1 (Dual1 (d t n m)) :~: Dual1 (d (Dual t) n m)
-lemma2 _ Refl = Refl
-
-lemma3 :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (d (Dual t) n m) :~: d (Dual (Dual t)) n m
-lemma3 d = lemma4 d Refl 
-
-lemma4 :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual t :~: t' -> Dual1 (d t' n m) :~: d (Dual t') n m
-lemma4 (DiagrammaticDuality _) = diagrammaticSoundDual1
-
-reflDual1'' :: DiagrammaticDuality s d i o (d t n m)
-  -> Dual1 (Dual1 (d t n m)) :~: Dual1 (d (Dual t) n m) -- lemma1
-  -> Dual1 (d (Dual t) n m) :~: d (Dual (Dual t)) n m   -- lemma3
-  -> Dual1 (Dual1 (d t n m)) :~: d t n m
-reflDual1'' (DiagrammaticDuality Refl) Refl Refl = Refl
-
-instance (TransformableTyp s, Transformable1 Op s)
-  => Duality1 s (DiagrammaticDuality s d) (IsoOp s) Op where
-  toDual1 d@(DiagrammaticDuality _)  = case dgmdSoundDual1 d of Refl -> coDiagrammatic
-  isoBidual1 (DiagrammaticDuality _) = Functor1 . isoToOpOpStruct   
--}
 --------------------------------------------------------------------------------
 -- prpDiagrammaticApplicative -
 
 relDiagrammaticApplicative :: (DiagrammaticApplicative h d, Show (d t n m a))
   => Struct Ort b -> h a b -> d t n m a -> Statement
 relDiagrammaticApplicative Struct h a
-  =  (diagram (dmap h a) == dgMap h (diagram a)) :?> Params ["a":=show a]
+  = (diagram (dmap h a) == dmap h (diagram a)) :?> Params ["a":=show a]
 
 -- | validity according to 'DiagrammaticApplicative'.
 prpDiagrammaticApplicative :: (DiagrammaticApplicative h d, Show (d t n m a))
@@ -213,6 +199,17 @@ prpDiagrammaticApplicative :: (DiagrammaticApplicative h d, Show (d t n m a))
 prpDiagrammaticApplicative h a = Prp "DiagrammaticApplicative" :<=>:
   relDiagrammaticApplicative (tau (range h)) h a
 
+--------------------------------------------------------------------------------
+-- prpDiagrammaticApplicative1 -
+
+-- | validity according to 'DiagrammaticApplicative1'.
+prpDiagrammaticApplicative1 :: ( DiagrammaticApplicative1 h d t n m
+                               , Eq (d t n m b), Show (d t n m a)
+                               )
+  => h a b -> d t n m a -> Statement
+prpDiagrammaticApplicative1 h d = Prp "DiagrammaticApplicative1" :<=>:
+  (dmap h d == amap1 h d) :?> Params ["d":= show d]
+  
 --------------------------------------------------------------------------------
 -- prpDiagrammaticDualisable -
 
@@ -225,14 +222,4 @@ prpDiagrammaticDualisable :: ( DiagrammaticDualisable s d
 prpDiagrammaticDualisable s Refl d = Prp "DiagrammaticDualisable" :<=>:
   (coDiagrammatic (tauOp s) (coDiagrammatic s d) == dmap (isoToOpOpStruct s) d)
     :?> Params ["d":=show d]
-
---------------------------------------------------------------------------------
--- dgmDuality -
-
-instance (Category h, HomOriented h) => Functorial1 h (Diagram t n m)
-
-dgmDuality :: (TransformableOrt s, TransformableOp s, TransformableTyp s)
-  => Dual (Dual t) :~: t
-  -> DiagrammaticDuality s Diagram (IsoOp s) Op (Diagram t n m) (Diagram (Dual t) n m)
-dgmDuality = DiagrammaticDuality
